Successful emittion of "Recent Posts" page.
