+++ quartz.config.ts
@@ -109,6 +109,7 @@ const config: QuartzConfig = {
       Plugin.Assets(),
       Plugin.Static(),
       Plugin.NotFoundPage(),
+      Plugin.RecentPosts(),
     ],
   },
 }
