+++ quartz/components/RecentNotes.tsx
@@ -8,7 +8,7 @@ import { GlobalConfiguration } from "../cfg"
 import { i18n } from "../i18n"
 import { classNames } from "../util/lang"
 
-interface Options {
+export interface Options {
   title?: string
   limit: number
   linkToMore: SimpleSlug | false
