+++ quartz/plugins/emitters/index.ts
@@ -8,3 +8,4 @@ export { Static } from "./static"
 export { ComponentResources } from "./componentResources"
 export { NotFoundPage } from "./404"
 export { CNAME } from "./cname"
+export { RecentPosts } from "./recentPosts"
